src/agents/claude_chatbot/requirements.txt
fastapi
uvicorn
boto3
mlflow
requests
pydantic
pytest
strands-agents
bedrock-agentcore-starter-toolkit
agentcore
pandas
langgraph
python-dotenv
langchain